requirements-startup.txt
Pillow>=5.3.0
PyGithub
aiofiles
spotdl
aiohttp
pygments
aria2p
async_generator
beautifulsoup4
bs4
bwb==3.0.0
cfscrape
coffeehouse
cowpy
cryptg
dnspython
emoji
gTTS-token>=1.1.3
gTTS>=2.0.1
gitpython
glitch_this
google-api-python-client==1.8.0
google-auth-oauthlib
google_images_download>=2.8.0
googletrans==3.1.0a0
geopy
gsearch
hachoir
heroku3
httplib2
humanize
justwatch
lxml
oauth2client
patool
psycopg2
psycopg2-binary
PyLyrics
pySmartDL
pybase64>=0.4.0
PyDictionary
pyfiglet
pylast
pymongo
python-barcode
python-dotenv
python-magic
pytube>=9.5.1
pytz
qrcode
regex
requests>=2.18.4
search-engine-parser>=0.4.2
selenium
speedtest-cli>=2.0.2
spotify_token
sqlalchemy>=1.2
telegraph
urbandict>=0.5
validators
wikipedia>=1.4.0
youtube-dl
covid
tswift
